LayoutTests/fast/images/text-recognition/image-overlay-text-without-leading-whitespace-expected.txt

PASS input.value is "foobar baz"
PASS successfullyParsed is true

TEST COMPLETE
